chore(folio): extract keyboard shortcuts into a hook
Move the document-level keydown effect from `DocxEditor.tsx` into
`hooks/useKeyboardShortcuts.ts`. The hook listens for:
 - Cmd/Ctrl+F → open find dialog with the current selection
 - Cmd/Ctrl+H → open replace dialog
 - Cmd/Ctrl+P → invoke the editor's custom print path
 - Delete/Backspace → delete a fully-selected table (CellSelection or
   layout-overlay selection)

The print logic itself (`handleDirectPrint`) stays in `DocxEditor.tsx`
for now — it depends on `containerRef`, the i18n `t`, and the local
toast helper; the hook takes it as `onDirectPrint`. The whole-table
deletion branch is pulled into a `tryDeleteSelectedTable` helper so
the keydown handler reads top-to-bottom instead of nested two levels
deep.

No behavior change.
